Removed old-style scripts and tutorials
