+++ README.md
@@ -16,8 +16,6 @@
     - [5.1.4. Run the script](#514-run-the-script)
     - [5.1.5. Saving at intermediate steps](#515-saving-at-intermediate-steps)
   - [5.2. Enhanced Sampling (PT-WTE)/ Experiment Directed simulation (EDS)](#52-enhanced-sampling-pt-wte-experiment-directed-simulation-eds)
-    - [5.2.1. Preparation](#521-preparation)
-    - [5.2.2. PT-WTE](#522-pt-wte)
 - [6. Restarting Simulations](#6-restarting-simulations)
 
 # 1. Release Notes
@@ -202,103 +200,6 @@ python simple.py
 ## 5.2. Enhanced Sampling (PT-WTE)/ Experiment Directed simulation (EDS)
 
 A complete example can be found in `peptidesim/inputs/pte`.
-
-### 5.2.1. Preparation
-Create a Python script for the preparation step. It is similiar to NVT simulation script.
-
-If you are using experimental bias, additional code will be needed for the preparation. Check part1.py, [plumed METAD documentation](https://www.plumed.org/doc-v2.5/user-doc/html/_m_e_t_a_d.html) and [plumed EDS documentation](https://www.plumed.org/doc-v2.5/user-doc/html/_e_d_s.html).
-
-### 5.2.2. PT-WTE
-
-Create a second script and import all required modules.
-Specify peptides, replica number and exchange period. Note that the name of the job must be consistant with that in part 1.
-
-```python
-name = nameinpart1
-MPI_NP = 16
-peptide_cpoies = 1 #number of peptide per replica
-replicas = 16 #number of replicas
-remd_exhcange_period = 250
-#reload
-ps=3
-```
-
-We will need a function for getting replica exchange efficiency. The function is used to extract replica exchange efficiency from log files, which will be used as a criteria for ending replica exchange iterations.
-```python
-def get_replex_e(ps, replica_number):
-    with open(ps.sims[-1].location + '/' + ps.sims[-1].metadata['md-log']) as f:
-        p1 = re.compile('Repl  average probabilities:')
-        p2 = re.compile('Repl\s*' + ''.join(['([0-9\.]+\s*)' for _ in range(replica_number - 1)]) + '$')
-        ready = False
-        answer=-1
-        for line in f.readlines():
-            if not ready and p1.findall(line):
-                ready = True
-            elif ready:
-                match = p2.match(line)
-
-                if match:
-                    answer= [float(s) for s in match.groups()]
-                    break
-
-        return answer
-```
-
-Specify conditions for replica exchange. The `ps.pte_replica` function generates plumed scripts based on the arguments inputted.
-
-```python
-pte_result = ps.pte_replica(mpi_np=MPI_NP, max_tries=3,min_iters=1, mdp_kwargs={'nsteps':int(400* 5*10**2), }, replicas=replicas,hills_file_location=os.getcwd(),hot=400,hill_height=0.6,sigma=250,bias_factor=16, eff_threshold=0.20,cold=278,exchange_period=200)
-pte_plumed_script=pte_result['plumed']
-replica_temps=pte_result['temperatures']
-with open('plumed_pte.dat', 'w') as f:
-    f.write(pte_plumed_script)
-ps.add_file('plumed_pte.dat')
-temps = ','.join(str(e) for e in replica_temps)
-kwargs = [ {'ref_t': ti} for ti in replica_temps]
-time_ns=5.0 # run time for each iteration in ns
-replex_eff = 0.2 # desired replica exchange efficiency
-for kw in kwargs:
-    kw['nsteps']=  int(time_ns * 5 * 10 ** 5)
-max_iterations=20 # maximum number of iteration allowed
-min_iterations=10 # minimum number of iteration before ending replica exchange
-```
-
-If you are running EDS or require additional output from plumed, you could add additional plumed script by
-```python
-plumed_input0=textwrap.dedent(
-    '''
-    plumed script
-    ''')
-plumed_input=pte_plumed_script+plumed_input0
-with open('plumed_eds_conver_pt_wte_metad.dat', 'w') as f:
-    f.write(plumed_input)
-ps.add_file('plumed_eds_conver_pt_wte_metad.dat')
-```
-
-We can now start replica exchange. Iteration will break if the replica exchange efficiency reaches desired value and minimum number of iteration is reached. Proceed to NVT_production step after replica exchange.
-```python
-for i in range(max_iterations):
-    ps.run(mdpfile='peptidesim_nvt.mdp', tag='nvt_conver_{}'.format(i),
-           mdp_kwargs=kwargs, run_kwargs={'plumed':'plumed.dat','replex':remd_exhcange_period}, mpi_np=MPI_NP)
-
-    rep_eff_1 = get_replex_e(ps, replicas)
-    if rep_eff_1 ==-1:
-        ps.run(mdpfile='peptidesim_nvt.mdp', tag='nvt_conver_{}'.format(i),  mdp_kwargs=kwargs,mpi_np=MPI_NP, run_kwargs={'plumed':'plumed.dat','replex':remd_exhcange_period},mpi_np=MPI_NP)
-
-    elif (min(rep_eff_1) >= replex_eff and i>=min_iterations):
-        print('Reached replica exchange efficiency of {}. Continuing to production'.format(rep_eff_1))
-        break
-    else:
-        print('Replica exchange efficiency of {}. Continuing simulation'.format(rep_eff_1))
-
-
-final_time_eds=int(0.040*5*10**5)
-ps.run(mdpfile='peptidesim_nvt.mdp', tag='nvt_prod',  mdp_kwargs={'nsteps': final_time_eds, 'ref_t': 278},mpi_np=MPI_NP)
-```
-
-Write a bash script and run.
-
-**Note**: During replica exchange simulations, MPI_NP must be consistant with N (number of nodes), ntask_per_node and c (cpus_per_task) in bash script. Restarts should also have the same number of MPI processes otherwise it will result in PTE tuning log file missing error.
 
 # 6. Restarting Simulations
