--- inputs/pte/example.json
@@ -1,1 +1,0 @@
-{"RE":["19,20,21,22,23,24","41,42,43,44,45"], "KE":["19,20,21,22","38,39,40,41,42,43"],"RD":["19,20,21,22,23,24","38,39,40,41,42"],"KD":["19,20,21,22","35,36,37,38,39,40"],"hill_height":1,"mpi_np":8,"replica_number":8,"hot_temperature":400,"max_iterations":25,"final_time":5}
